+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Icon: better `dark` theme on icon with shape.
+
 ## [2.2.11][] - 2022-03-17
 
 ### Fixed